Initial Run

+++ install.sh
@@ -18,21 +18,29 @@ brew bundle
 chsh -s $(which zsh)
 
 # Install Composer
-curl -sS https://getcomposer.org/installer | php
-mv composer.phar /usr/local/bin/composer
+if test ! $(which composer)
+then
+  echo "  Installing Composer for you."
+
+    php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+    php -r "if (hash_file('SHA384', 'composer-setup.php') === '55d6ead61b29c7bdee5cccfb50076874187bd9f21f65d8991d46ec5cc90518f447387fb9f76ebae1fbbacf329e583e30') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
+    php composer-setup.php
+    php -r "unlink('composer-setup.php');"
+fi
+
+if [ -f composer.phar ];
+  mv composer.phar /usr/local/bin/composer
+fi
 
 # Install global Composer packages
-/usr/local/bin/composer global require laravel/installer laravel/lumen-installer laravel/valet tightenco/jigsaw spatie/http-status-check bramus/mixed-content-scan laravel/spark-installer
+/usr/local/bin/composer global require laravel/installer laravel/lumen-installer tightenco/jigsaw spatie/http-status-check
 
 # Install Laravel Valet
-$HOME/.composer/vendor/bin/valet install
-
-# Install global NPM packages
-npm install --global yarn
+#$HOME/.composer/vendor/bin/valet install
 
 # Create a Sites directory
 # This is a default directory for macOS user accounts but doesn't comes pre-installed
-mkdir $HOME/Sites
+mkdir $HOME/Operations
 
 # Set macOS preferences
 # We will run this last because this will reload the shell
